Immutable Data: All the Shapes and Sizes
Persistent data structures are a great fit for React; replacing
mutable arrays and objects with immutable vectors and hash maps
can lead to hugely compelling benefits in terms of robustness
(somewhat expected) and performance (who expected that?). This has
been demonstrated in a compile-to-JavaScript setting by
ClojureScript and Om, and in pure JavaScript by Immutable.js; see
Lee Byron's React.js Conf talk for an introduction to the topic.

The landscape of persistent data structures is vastly richer,
however, and this talk aims to explore in search of specimens of
interest to web developers today. In particular, we will take a
deep dive into the key implementation ideas behind the following:

  - transients, enabling efficient "batch updates" to persistent
    data structures (rivalling native mutable built-ins in
    performance) with no harmful impact to the base PDS semantics;

  - RRB trees, bringing efficient concatenation and slicing to
    persistent vectors;

  - binary search trees annotated with extra information for
    sorted dictionaries and sets with logarithmic-time extraction
    of subsequences and splicing of instances with non-overlapping
    key ranges;

  - and some more exotic friends and relations of the above.