How React & Flux Turn Applications Into Extensible Platforms
https://www.youtube.com/watch?v=rccmUc1bQ4o

Chrome is great, but 3rd party extensions make it better. The iPhone
is great, but apps make it better. You React-app may be great, but
imagine if you could safely and robustly allow 3rd party extensions to
enhance it.

We'll talk about specific features of React & Flux, React CSS,
programming design patterns, and custom libraries, which can turn a
static application into a dynamic platform that an ecosystem of
developers can build on top of.

We've built a highly-extensible desktop email client with React & Flux
on Atom Shell, and we'll also show concrete examples of where these
tools enabled a 3rd party ecosystem of email plugins.

Our goal is for you to take away how to use React to be more than just
great application developers, but now great platform developers as
well.